+++ packages/tldraw/src/lib/shapes/note/NoteShapeUtil.tsx
@@ -36,6 +36,7 @@ import {
 } from '../shared/default-shape-constants'
 import { getFontDefForExport } from '../shared/defaultStyleDefs'
 
+import { startEditingShapeWithLabel } from '../shared/TextHelpers'
 import { useForceSolid } from '../shared/useForceSolid'
 import {
 	ADJACENT_NOTE_MARGIN,
@@ -43,7 +44,6 @@ import {
 	NOTE_CENTER_OFFSET,
 	NOTE_SIZE,
 	getNoteShapeForAdjacentPosition,
-	startEditingNoteShape,
 } from './noteHelpers'
 
 /** @public */
@@ -450,7 +450,7 @@ function useNoteKeydownHandler(id: TLShapeId) {
 				const newNote = getNoteShapeForAdjacentPosition(editor, shape, adjacentCenter, pageRotation)
 
 				if (newNote) {
-					startEditingNoteShape(editor, newNote)
+					startEditingShapeWithLabel(editor, newNote, true /* selectAll */)
 				}
 			}
 		},
